lab2/requirements.txt
numpy
pandas
streamlit
python-dotenv